# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Provenance {
  source: string;
  data_type: string;
  confidence: string;
}

export type StudyAreaBounds = {
  min_lat: number;
  max_lat: number;
  min_lon: number;
  max_lon: number;
};

export interface DashboardMetrics {
  power: number;
  water: number;
  roads: number;
  telecom: number;
  population: number;
  assets: number;
  buildings: number;
  roads_count: number;
}

export interface StudyArea {
  name: string;
  city: string;
  state: string;
  country: string;
  bounds: StudyAreaBounds;
  metrics: DashboardMetrics;
}

export interface Point {
  lat: number;
  lon: number;
}

export interface Asset {
  id: string;
  type: string;
  name: string;
  location: Point;
  capacity: string;
  importance: number;
  status: string;
  criticality: number;
  dependencies: number;
  provenance: Provenance;
}

export interface Building {
  id: string;
  building_type: string;
  centroid: Point;
  area_m2: number;
  levels: number;
  height: number;
  provenance: Provenance;
  geometry?: number[][][];
}

export interface Road {
  id: string;
  name: string;
  road_type: string;
  geometry: Point[];
  length_m: number;
  provenance: Provenance;
}

export interface Dependency {
  source_id: string;
  target_id: string;
  dependency_type: string;
  strength: number;
  threshold: number;
  confidence: string;
}

export interface DependencyGraph {
  nodes: Asset[];
  edges: Dependency[];
}

export interface TwinData {
  buildings: Building[];
  roads: Road[];
  assets: Asset[];
  dependencies: Dependency[];
}

export type SimulationInputAction = typeof SimulationInputAction[keyof typeof SimulationInputAction];


export const SimulationInputAction = {
  FAIL: 'FAIL',
  DEGRADE: 'DEGRADE',
  DISCONNECT: 'DISCONNECT',
} as const;

export interface SimulationInput {
  asset_id: string;
  action: SimulationInputAction;
  duration_minutes?: number;
  interventions?: string[];
}

export interface TimelineEvent {
  time: number;
  label: string;
  detail: string;
  asset_id: string;
  state: string;
}

export interface ImpactMetrics {
  population_affected: number;
  assets_affected: number;
  failed_assets: number;
  degraded_assets: number;
  critical_services: number;
  cascade_depth: number;
  area_affected_km2: number;
  roads_affected: number;
  buildings_affected: number;
  dependency_edges_affected: number;
}

export interface SimulationResult {
  id: string;
  scenario: string;
  source_asset_id: string;
  impact: ImpactMetrics;
  timeline: TimelineEvent[];
  affected_asset_ids: string[];
  explanation: string;
}

export interface Criticality {
  asset_id: string;
  name: string;
  type: string;
  score: number;
  reasons: string[];
}

export interface OptimizeInput {
  /** @minimum 0 */
  budget_crore: number;
}

export interface Intervention {
  id: string;
  name: string;
  cost_crore: number;
  target: string;
  impact_reduction: number;
  implementation_time: string;
  rationale: string;
}

export interface OptimizationResult {
  budget_crore: number;
  total_cost_crore: number;
  estimated_impact_reduction: number;
  interventions: Intervention[];
  method: string;
}

export interface CopilotInput {
  /** @minLength 1 */
  question: string;
}

/**
 * @nullable
 */
export type CopilotResultSimulation = { [key: string]: unknown } | null;

export interface CopilotResult {
  answer: string;
  intent: string;
  /** @nullable */
  simulation: CopilotResultSimulation;
}

export type ListAssetsParams = {
type?: string;
status?: string;
};