# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
import type { Point } from './point';
import type { Provenance } from './provenance';

export interface Asset {
  id: string;
  type: string;
  name: string;
  location: Point;
  capacity: string;
  importance: number;
  status: string;
  criticality: number;
  dependencies: number;
  provenance: Provenance;
}