# lib/api-zod/src/generated/types/copilotResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
import type { CopilotResultSimulation } from './copilotResultSimulation';

export interface CopilotResult {
  answer: string;
  intent: string;
  /** @nullable */
  simulation: CopilotResultSimulation;
}