# lib/api-zod/src/generated/types/dependency.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */

export interface Dependency {
  source_id: string;
  target_id: string;
  dependency_type: string;
  strength: number;
  threshold: number;
  confidence: string;
}