# lib/api-zod/src/generated/types/optimizationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
import type { Intervention } from './intervention';

export interface OptimizationResult {
  budget_crore: number;
  total_cost_crore: number;
  estimated_impact_reduction: number;
  interventions: Intervention[];
  method: string;
}