# lib/api-zod/src/generated/types/optimizeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */

export interface OptimizeInput {
  /** @minimum 0 */
  budget_crore: number;
}