# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationInputAction } from './simulationInputAction';

export interface SimulationInput {
  asset_id: string;
  action: SimulationInputAction;
  duration_minutes?: number;
  interventions?: string[];
}