# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  time: number;
  label: string;
  detail: string;
  asset_id: string;
  state: string;
}