# lib/api-zod/src/generated/types/twinData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RIPPLE urban infrastructure resilience digital twin API
 * OpenAPI spec version: 0.1.0
 */
import type { Asset } from './asset';
import type { Building } from './building';
import type { Dependency } from './dependency';
import type { Road } from './road';

export interface TwinData {
  buildings: Building[];
  roads: Road[];
  assets: Asset[];
  dependencies: Dependency[];
}